upates
The website has received a total of 10 signups for the giveaway. The only thing that is left is to make it responsive in landscape on mobile and perhaps to incorporate some of the design from the pinterest website layout that was provided to me onto the website. Otherwise, I think it's pretty good and otherwise have expressed the same sentiment. As in the words of my friend Javid Wilson "Well done!"